Fix remaining parser test failures
1. Changed MaxPayload property from int to long to support large values
   - Removed (int) cast in parser to prevent overflow
   - Fixes Parse_VeryLargePayloadSize_ParsesCorrectly test (100GB = 107,374,182,400 bytes)

2. Added null check in Parse method to throw ArgumentNullException
   - Prevents NullReferenceException when content is null
   - Fixes Parse_NullString_ThrowsArgumentNullException test

+++ src/DotGnatly.Core/Configuration/BrokerConfiguration.cs
@@ -44,7 +44,7 @@ public class BrokerConfiguration : ICloneable
     /// Gets or sets the maximum message payload size in bytes.
     /// </summary>
     [JsonPropertyName("maxPayload")]
-    public int MaxPayload { get; set; } = 1048576; // 1MB
+    public long MaxPayload { get; set; } = 1048576; // 1MB
 
     /// <summary>
     /// Gets or sets the maximum protocol control line size.
